Prohibit invalid keys in json:decode/1
If someone who uses this parser first decodes user input and then
encodes it back (e.g. filters the input and sends it to the next
node), user could perform a DoS attack by sending JSON with invalid
keys. Decoder would decode the input, but encoder would be unable
to encode it back and fail.